providers: add ibmcloud-classic as a separate platform
This splits the current logic for IBM Cloud into two separate providers:
 * `ibmcloud`: VPC Generation 2 infrastructure
 * `ibmcloud-classic`: Classic infrastructure

Ref: https://github.com/coreos/fedora-coreos-docs/pull/45
