+++ src/metadata.rs
@@ -21,7 +21,8 @@ use crate::providers::cloudstack::configdrive::ConfigDrive;
 use crate::providers::cloudstack::network::CloudstackNetwork;
 use crate::providers::digitalocean::DigitalOceanProvider;
 use crate::providers::gcp::GcpProvider;
-use crate::providers::ibmcloud;
+use crate::providers::ibmcloud::IBMGen2Provider;
+use crate::providers::ibmcloud_classic::IBMClassicProvider;
 use crate::providers::openstack::network::OpenstackProvider;
 use crate::providers::packet::PacketProvider;
 use crate::providers::vagrant_virtualbox::VagrantVirtualboxProvider;
@@ -52,8 +53,10 @@ pub fn fetch_metadata(provider: &str) -> errors::Result<Box<dyn providers::Metad
         "gce" => box_result!(GcpProvider::try_new()?),
         #[cfg(not(feature = "cl-legacy"))]
         "gcp" => box_result!(GcpProvider::try_new()?),
-        // ibmcloud has multiple infrastructure types, so this internally tries to auto-detect.
-        "ibmcloud" => ibmcloud::try_autodetect(),
+        // IBM Cloud - VPC Generation 2.
+        "ibmcloud" => box_result!(IBMGen2Provider::try_new()?),
+        // IBM Cloud - Classic infrastructure.
+        "ibmcloud-classic" => box_result!(IBMClassicProvider::try_new()?),
         "openstack-metadata" => box_result!(OpenstackProvider::try_new()?),
         "packet" => box_result!(PacketProvider::try_new()?),
         "vagrant-virtualbox" => box_result!(VagrantVirtualboxProvider::new()),
