+++ src/providers/mod.rs
@@ -30,6 +30,7 @@ pub mod cloudstack;
 pub mod digitalocean;
 pub mod gcp;
 pub mod ibmcloud;
+pub mod ibmcloud_classic;
 pub mod openstack;
 pub mod packet;
 pub mod vagrant_virtualbox;
