+++ tests/fixtures/ibmcloud-classic/meta_data.json
