+++ tests/fixtures/ibmcloud-classic/network_data.json
